refactor: MimeTypeResolver 패키지 이동
